requirements.txt
attrs>=22.2.0
Automat>=22.10.0
beautifulsoup4>=4.9.3
bs4>=0.0.1
certifi>=2022.12.7
cffi>=1.15.1
chardet>=4.0.0
charset-normalizer>=3.1.0
click>=8.1.3
constantly>=15.1.0
cramjam>=2.6.2
cryptography>=39.0.2
cssselect>=1.2.0
dnspython>=2.3.0
fastparquet>=2023.2.0
filelock>=3.9.0
fsspec>=2023.3.0
googlesearch-python>=1.1.0
hyperlink>=21.0.0
idna>=2.10
incremental>=22.10.0
itemadapter>=0.7.0
itemloaders>=1.0.6
jmespath>=1.0.1
lxml>=4.9.2
numpy>=1.24.2
packaging>=23.0
pandas>=1.5.3
parsel>=1.7.0
Protego>=0.2.1
pyarrow>=11.0.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.21
PyDispatcher>=2.0.7
pymongo>=4.3.3
pyOpenSSL>=23.0.0
python-dateutil>=2.8.2
pytz>=2022.7.1
queuelib>=1.6.2
requests>=2.28.2
requests-file>=1.5.1
scr>=1.0.0
Scrapy>=2.8.0
service-identity>=21.1.0
shapely>=2.0.1
six>=1.16.0
soupsieve>=2.4
tldextract>=3.4.0
Twisted>=22.10.0
typing_extensions>=4.5.0
urllib3>=1.26.15
w3lib>=2.1.1
zope.interface>=5.5.2
